init: preparing the repository for the start of the task
