+++ .env.example
@@ -1,0 +1,2 @@
+HTTP_PORT = 3000
+WSS_PORT = 8080
