+++ .gitignore
@@ -14,3 +14,7 @@ Thumbs.db
 
 # Logs
 npm-debug.log
+
+.env
+
+/dist
