+++ package.json
@@ -3,20 +3,22 @@
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
     "main": "index.js",
-    "type": "module",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
+        "start:dev": "nodemon dist/index.js",
+        "start:prod": "tsc & nodemon dist/index.js",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "author": "Andrei Auchynnikau",
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
+        "@types/ws": "^8.5.4",
         "nodemon": "^2.0.16"
     },
     "dependencies": {
         "@nut-tree/nut-js": "^2.3.0",
+        "dotenv": "^16.0.3",
+        "jimp": "^0.16.2",
         "ws": "^8.8.0"
     }
 }
